fix: add edit expense and delete

+++ frontend/app/src/api/expense.ts
@@ -2,8 +2,10 @@ import { i18n } from "@lingui/core";
 import { msg } from "@lingui/core/macro";
 import { mutationOptions } from "@tanstack/react-query";
 import z from "zod";
+import { Alert } from "@/components/AlertRoot";
 import { Snackbar } from "@/components/SnackbarRoot";
 import { fetchJSON, fetchRaw } from "@/query/fetch";
+import type { Group } from "./group";
 
 export const UpsertExpense = z.object({
 	expense_name: z.string(),
@@ -87,24 +89,58 @@ export function updateExpense() {
 
 const deleteExpenseSuccessTitle = msg`Expense deleted`;
 const deleteExpenseFailedTitle = msg`Failed to delete expense, something went wrong`;
+const deleteExpenseDisclaimerTitle = msg`Delete expense?`;
+const deleteExpenseDisclaimerBody = msg`Are you sure you want to delete the expense {name}`;
+const deleteExpenseDisclaimerCancel = msg`Cancel`;
+const deleteExpenseDisclaimerDelete = msg`Delete`;
+
 type DeleteExpenseArguments = {
 	groupId: number;
 	expenseId: number;
 };
 export function deleteExpense() {
 	return mutationOptions({
-		async mutationFn({ groupId, expenseId }: DeleteExpenseArguments) {
-			return await fetchRaw(`groups/${groupId}/expense/${expenseId}`, {
+		async mutationFn({ groupId, expenseId }: DeleteExpenseArguments, context) {
+			const group = context.client.getQueryData<Group>([
+				"groups",
+				{ id: groupId.toString() },
+			]);
+
+			const expense = group?.expenses.find(
+				(expense) => expense.id === expenseId,
+			);
+			deleteExpenseDisclaimerBody.values = {
+				name: expense?.name,
+			};
+			console.log(deleteExpenseDisclaimerBody);
+			const result = await Alert.destructive({
+				cancelText: i18n._(deleteExpenseDisclaimerCancel),
+				primaryText: i18n._(deleteExpenseDisclaimerDelete),
+				title: i18n._(deleteExpenseDisclaimerTitle),
+				body: i18n._(deleteExpenseDisclaimerBody),
+			});
+			if (result !== "success") {
+				return {
+					result,
+				};
+			}
+
+			await fetchRaw(`groups/${groupId}/expense/${expenseId}`, {
 				method: "DELETE",
 			});
+			return {
+				result,
+			};
 		},
 		onSuccess(data, variables, onMutateResult, context) {
-			context.client.invalidateQueries({
-				queryKey: ["groups"],
-			});
-			Snackbar.toast({
-				text: i18n._(deleteExpenseSuccessTitle),
-			});
+			if (data.result === "success") {
+				context.client.invalidateQueries({
+					queryKey: ["groups"],
+				});
+				Snackbar.toast({
+					text: i18n._(deleteExpenseSuccessTitle),
+				});
+			}
 		},
 		onError(error, variables, onMutateResult, context) {
 			Snackbar.toast({
